add tests

+++ model/domain/category.go
@@ -1,6 +1,6 @@
 package domain
 
 type Category struct {
-	Id   int    `gorm:"primary_key; column:id"`
+	Id   uint64 `gorm:"primary_key; column:id"`
 	Name string `gorm:"column:name"`
 }
