+++ model/web/category_update_request.go
@@ -1,6 +1,6 @@
 package web
 
 type CategoryUpdateRequest struct {
-	Id   int    `validate:"required"`
+	Id   uint64 `validate:"required"`
 	Name string `validate:"required,max=200,min=1" json:"name"`
 }
