+++ repository/category_repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"context"
+
 	"github.com/aronipurwanto/go-restful-api/model/domain"
 )
 
@@ -9,6 +10,6 @@ type CategoryRepository interface {
 	Save(ctx context.Context, category domain.Category) (domain.Category, error)
 	Update(ctx context.Context, category domain.Category) (domain.Category, error)
 	Delete(ctx context.Context, category domain.Category) error
-	FindById(ctx context.Context, categoryId int) (domain.Category, error)
+	FindById(ctx context.Context, categoryId uint64) (domain.Category, error)
 	FindAll(ctx context.Context) ([]domain.Category, error)
 }
